requirements.txt
yagmail
pandas
requests
python-dotenv
openpyxl

